Added function to get a PID from a process name
With this function, it is possible to get the PID from a process name.
The goal would be to allow the user to send a process name and the
daemon gets the name and tries to identify the PID.
Ideally, if it finds more than one, the daemon should ask which one.
It should also offer the option to bypass all that by giving directly a
PID.